refactor: rename `AppContext` => `SlidesContext`
